api/requirements.txt
pydantic-settings==2.9.1
fastapi[standard]==0.115.12
strawberry-graphql[fastapi]==0.270.2
alembic==1.16.1
sqlalchemy==2.0.41
asyncpg==0.30.0
greenlet==3.2.2
